refactor: call kleros scout directly
